+++ packages/vtable/src/tools/update-row.ts
@@ -11,8 +11,8 @@ import type { BaseTableAPI } from '../ts-types/base-table';
  *   2.2 row inside of proxy row range => update row to proxy rowEnd
  * REM:
  * 1. row before rowStart or after rowEnd => skip
- * 2. no row after rowEnd => skip
- * 3. rows after rowEnd => update row to proxy rowEnd
+ * 2. row after rowStart and before new rowEnd => update row to rowEnd
+ * 3. rows after new rowEnd before old rowEnd => delete
  */
 
 export function fixUpdateRowRange(
@@ -32,7 +32,7 @@ export function fixUpdateRowRange(
   const { addCellPositions, removeCellPositions } = diffPositions;
   const proxy = table.scenegraph.proxy;
   const { rowStart, rowLimit } = proxy;
-  let { rowEnd, bodyBottomRow } = proxy;
+  let { rowEnd } = proxy;
   let updateRow = Infinity;
 
   for (let i = 0; i < addCellPositions.length; i++) {
@@ -53,20 +53,22 @@ export function fixUpdateRowRange(
     }
   }
 
+  const newRowEnd = Math.min(rowStart + rowLimit, table.rowCount - 1 - table.bottomFrozenRowCount);
   for (let i = 0; i < removeCellPositions.length; i++) {
     const { row: cellRow } = removeCellPositions[removeCellPositions.length - i - 1];
     if (cellRow < rowStart || cellRow > rowEnd) {
-      // continue;
-      bodyBottomRow--;
-    } else if (rowEnd === bodyBottomRow) {
+      // row before rowStart or after rowEnd => skip
+      continue;
+    } else if (cellRow > newRowEnd) {
+      // rows after new rowEnd before old rowEnd => delete
       removeCells.push({
         col,
         row: cellRow
       });
       updateRow--;
     } else {
+      // row after rowStart and before new rowEnd => update row to rowEnd
       updateRow = Math.min(updateRow, cellRow);
-      bodyBottomRow--;
     }
   }